Refactor with send_email in utils. Reference #37
